Task2_Senior_Citizen_Identification/requirements.txt
numpy
pandas
opencv-python
tensorflow
keras
imutils
datetime
streamlit
Pillow